requirements.txt
python-telegram-bot==20.0
pymongo==4.0.1
solana==0.30.2
python-dotenv==0.19.2